# [UI] Wrap bagged metrics/params in their own component + improve dropdown perf