--- a/mlflow/server/js/src/components/ExperimentRunsTableCompactView.js
+++ b/mlflow/server/js/src/components/ExperimentRunsTableCompactView.js
@@ -37,7 +37,6 @@ const styles = {
 class ExperimentRunsTableCompactView extends Component {
   constructor(props) {
     super(props);
-    this.onHover = this.onHover.bind(this);
     this.getRow = this.getRow.bind(this);
   }
 
@@ -78,10 +77,6 @@ class ExperimentRunsTableCompactView extends Component {
     hoverState: {isMetric: false, isParam: false, key: ""},
   };
 
-  onHover({isParam, isMetric, key}) {
-    this.setState({ hoverState: {isParam, isMetric, key} });
-  }
-
   /** Returns a row of table content (i.e. a non-header row) corresponding to a single run. */
   getRow({ idx, isParent, hasExpander, expanderOpen, childrenIds }) {
     const {
@@ -101,7 +96,6 @@ class ExperimentRunsTableCompactView extends Component {
       unbaggedParams,
       onRemoveBagged,
     } = this.props;
-    const hoverState = this.state.hoverState;
     const runInfo = runInfos[idx];
     const paramsMap = ExperimentViewUtil.toParamsMap(paramsList[idx]);
     const metricsMap = ExperimentViewUtil.toMetricsMap(metricsList[idx]);
@@ -127,14 +121,15 @@ class ExperimentRunsTableCompactView extends Component {
     });
     // Add bagged params
     const paramsCellContents = baggedParams.map((paramKey) => {
-      const isHovered = hoverState.isParam && hoverState.key === paramKey;
       const keyname = "param-" + paramKey;
       const sortIcon = ExperimentViewUtil.getSortIcon(sortState, false, true, paramKey);
       return (<BaggedCell
         key={keyname}
         sortIcon={sortIcon}
-        keyName={paramKey} value={paramsMap[paramKey].getValue()} onHover={this.onHover}
-        setSortByHandler={setSortByHandler} isMetric={false} isParam isHovered={isHovered}
+        keyName={paramKey} value={paramsMap[paramKey].getValue()}
+        setSortByHandler={setSortByHandler}
+        isMetric={false}
+        isParam
         onRemoveBagged={onRemoveBagged}/>);
     });
     if (this.shouldShowBaggedColumn(true)) {
@@ -153,18 +148,15 @@ class ExperimentRunsTableCompactView extends Component {
     // Add bagged metrics
     const metricsCellContents = baggedMetrics.map((metricKey) => {
       const keyname = "metric-" + metricKey;
-      const isHovered = hoverState.isMetric && hoverState.key === metricKey;
       const sortIcon = ExperimentViewUtil.getSortIcon(sortState, true, false, metricKey);
       return (
         <BaggedCell key={keyname}
                     keyName={metricKey}
                     value={metricsMap[metricKey].getValue().toString()}
-                    onHover={this.onHover}
                     sortIcon={sortIcon}
                     setSortByHandler={setSortByHandler}
                     isMetric
                     isParam={false}
-                    isHovered={isHovered}
                     onRemoveBagged={onRemoveBagged}/>
       );
     });