This PR contains part of #745  and improves performance of the compact runs table by:

1) Wrapping bagged metric & param key-value pairs in their own `BaggedCell` component (allows React to avoid re-computing the virtual DOM elements corresponding to a key-value pair each time the table is rendered). 
2) Using a custom dropdown menu that only renders its children when the dropdown is displayed, as described in https://github.com/react-bootstrap/react-bootstrap/issues/2056.
3) Removing the highlight-on-hover behavior of metric/param key-value pairs.

## Performance Impact
See [ChromePerformanceProfiles.zip](https://github.com/mlflow/mlflow/files/2656256/ChromePerformanceProfiles.zip), which contains the following Chrome performance profiles on an example experiment with 30 runs, each with 100 metrics and 100 params:

### Master
1) WithoutBaggedCell30RunsSort-0fa229c743e482786ad6d42b96f6b07f4bd6749e: sorting on the date column within the test experiment takes ~3.5 seconds (3 seconds scripting, 0.5 seconds rendering
2) WithoutBaggedCell30Runs-0fa229c743e482786ad6d42b96f6b07f4bd6749e: initial page load, takes ~5.5 seconds (4.8 seconds scripting and 0.7 seconds rendering)

### This branch
3) WithBaggedCell30RunsSort-285afab4042a0e53a90a18442b3b035634493317: sorting on the date column takes ~1 second (~0.6 seconds scripting, 0.4 seconds rendering)
4) WithBaggedCell30RunsPageLoad-285afab4042a0e53a90a18442b3b035634493317:  initial page load, takes 3 seconds, (2.25 seconds scripting and 0.6 seconds rendering).

In addition, hovering over an individual metric/param is quite a bit faster.